add firebase support
sorry, another huge commit.

* add another storage backend that uses firebase.com; what firebase
  application is based on a user setting that is still kept in
  localstorage
* tested and rejected the firebase angular lib, built a small wrapper
  around their JS object using `$q` that doesn't need a bunch of manual
  `$scope.$apply()` crap
* use `uuid` to generate ids for firebase; their `push` method is a
  little cumbersome compared to `uuid.v4()`
* add a "remove" option for chore activity
* connect to firebase on startup if we're configured to do so

I'm not happy with the factory method in `chores.js` to figure out which
backend to use, I'd like to get that into DI, but I'm not sure if that's
better as a routing `resolve` entry or add it to settings. I think I can
also reduce the difference between firebase and localstorage persistence
by using a common data format.